+++ inst/examples/sandbox/helpers.R
@@ -1,7 +1,8 @@
 # some sample R expressions to try
 examples <- c(
   'toggle("test")',
-  'toggle(delay = 0.5, selector = "a")',
+  'delay(500, toggle("test"))',
+  'toggle(selector = "a")',
   'toggle("test", TRUE, "fade", 2)',
   'toggle(id = "test", anim = TRUE, time = 2, animType = "slide")',
   'hide("test")',